players sprite now changes direction

+++ Character.pde
@@ -1,7 +1,7 @@
 public abstract class Character {
   int x, y, hearts, size;
   boolean tint;
-  long time, time2;
+  double time, time2;
   public Character() {
   }
 }
